Design(#10): 마이페이지 컴포넌트화 & 나의 활동 관련 라우터 설정
