+++ src/app/mypage/bookmark/page.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const BookmarkPage = () => {
+  return (
+    <div>
+      즐겨찾기 페이지
+    </div>
+  );
+};
+
+export default BookmarkPage;
