+++ src/app/mypage/myreviews/page.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const MyReviews = () => {
+  return (
+    <div>
+      내가 쓴 리뷰 페이지
+    </div>
+  );
+};
+
+export default MyReviews;
