+++ src/components/mypage/SectionHeader/SectionHeader.jsx
@@ -1,0 +1,20 @@
+"use client";
+
+import React from 'react';
+import styled from 'styled-components';
+
+const SectionHeader = ({header}) => {
+  return (
+    <Container>
+      {header}
+    </Container>
+  );
+};
+
+export default SectionHeader;
+
+const Container = styled.div`
+  font-weight: bold;
+  margin-bottom: 10px;
+  width: 540px;
+`;
